week3/week3-final-project/services/order-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.3
redis==5.0.1
httpx==0.27.0
